requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
openpyxl
scikit-learn
plotly
python-dotenv